requirements.txt
flask
flask-sqlalchemy
flask-security-too
matplotlib
email-validator
bcrypt
celery
redis
secure-smtplib
